MD should now be migrated to using 64-bit integers for timesteps, seeing as how ~2 billion timesteps (maximum signed 32-bit integer) is actually within the realm of reason these days.

+++ modules/algorithms/src/main/java/ffx/algorithms/cli/DynamicsOptions.java
@@ -114,7 +114,7 @@ public class DynamicsOptions {
      */
     @Option(names = {"-n", "--numberOfSteps"}, paramLabel = "1000000",
             description = "Number of molecular dynamics steps.")
-    int steps = 1000000;
+    long steps = 1000000;
 
     /**
      * -o or --optimize saves low-energy snapshots discovered (only for single topology simulations).
